chore: use VampireDownedServerEventSystem for the kill feed
