+++ killfeed/VampireDownedServerEventSystemPatches.cs
@@ -1,0 +1,68 @@
+﻿using System;
+using System.Collections.Generic;
+using Bloodstone.API;
+using HarmonyLib;
+using ProjectM;
+using Unity.Collections;
+using Unity.Entities;
+using v_rising_discord_bot_companion.game;
+
+namespace v_rising_discord_bot_companion.killfeed;
+
+[HarmonyPatch(typeof(VampireDownedServerEventSystem))]
+public class VampireDownedServerEventSystemPatches {
+
+    private readonly static List<PvpKill> _pvpKills = new();
+
+    public static List<PvpKill> getPvpKills() {
+        removeExpiredPvpKills();
+        return _pvpKills;
+    }
+
+    private static void removeExpiredPvpKills() {
+        _pvpKills.RemoveAll(pvpKill => DateTime.UtcNow > pvpKill.Occurred.AddMinutes(10));
+    }
+
+    [HarmonyPostfix]
+    [HarmonyPatch("OnUpdate")]
+    public static void Postfix(VampireDownedServerEventSystem __instance) {
+
+        if (__instance.__OnUpdate_LambdaJob0_entityQuery == null) {
+            return;
+        }
+
+        var entities = __instance.__OnUpdate_LambdaJob0_entityQuery.ToEntityArray(Allocator.Temp);
+        foreach (var entity in entities) {
+            handleDownedEntity(entity);
+        }
+
+        removeExpiredPvpKills();
+    }
+
+    private static void handleDownedEntity(Entity entity) {
+
+        VampireDownedServerEventSystem.TryFindRootOwner(entity, 1, VWorld.Server.EntityManager, out var victimEntity);
+        var source = VWorld.Server.EntityManager.GetComponentData<VampireDownedBuff>(entity).Source;
+        VampireDownedServerEventSystem.TryFindRootOwner(source, 1, VWorld.Server.EntityManager, out var killerEntity);
+
+        if (!VWorld.Server.EntityManager.HasComponent<PlayerCharacter>(killerEntity)
+            || !VWorld.Server.EntityManager.HasComponent<PlayerCharacter>(victimEntity) || victimEntity.Equals(killerEntity)) {
+            return;
+        }
+
+        var killer = VCharacter.from(killerEntity);
+        var victim = VCharacter.from(victimEntity);
+
+        _pvpKills.Add(new PvpKill(
+            Killer: new Player(
+                Name: killer.Character.Name.ToString(),
+                GearLevel: killer.getGearLevel()
+            ),
+            Victim: new Player(
+                Name: victim.Character.Name.ToString(),
+                GearLevel: victim.getGearLevel()
+            ),
+            Occurred: DateTime.UtcNow
+        ));
+    }
+}
